Adding a bulk copy using the psql command-line tool

+++ src/jam/sql/SQLDb.java
@@ -84,7 +84,8 @@ public abstract class SQLDb {
             throw JamException.runtime(ex);
         }
         finally {
-            bulkFile.delete();
+            if (bulkFile != null)
+                bulkFile.delete();
         }
     }
